Exchange Rate & Rate providers merged into one provider
