+++ src/components/exchange-rates/exchange-rates.ts
@@ -7,8 +7,8 @@ import { Coin } from '../../providers/currency/currency';
 import {
   DateRanges,
   ExchangeRate,
-  ExchangeRatesProvider
-} from '../../providers/exchange-rates/exchange-rates';
+  RateProvider
+} from '../../providers/rate/rate';
 
 export interface Card {
   unitCode: string;
@@ -26,8 +26,8 @@ export interface Card {
   templateUrl: 'exchange-rates.html'
 })
 export class ExchangeRates {
-  public isIsoCodeSupported: boolean;
-  public isoCode: string;
+  public isFiatIsoCodeSupported: boolean;
+  public fiatIsoCode: string;
   public coins = [];
   public fiatCodes = [
     'USD',
@@ -45,7 +45,7 @@ export class ExchangeRates {
   constructor(
     private navCtrl: NavController,
     private currencyProvider: CurrencyProvider,
-    private exchangeRatesProvider: ExchangeRatesProvider,
+    private rateProvider: RateProvider,
     private configProvider: ConfigProvider,
     private logger: Logger,
     private events: Events
@@ -83,8 +83,8 @@ export class ExchangeRates {
 
     // TODO: Add a new endpoint in BWS that
     // provides JUST  the current prices and the delta.
-    this.exchangeRatesProvider
-      .fetchHistoricalRates(this.isoCode, force, DateRanges.Day)
+    this.rateProvider
+      .fetchHistoricalRates(this.fiatIsoCode, force, DateRanges.Day)
       .then(response => {
         _.forEach(this.coins, (coin, index) => {
           if (response[coin.unitCode])
@@ -112,8 +112,11 @@ export class ExchangeRates {
   private setIsoCode() {
     const alternativeIsoCode = this.configProvider.get().wallet.settings
       .alternativeIsoCode;
-    this.isIsoCodeSupported = _.includes(this.fiatCodes, alternativeIsoCode);
-    this.isoCode = this.isIsoCodeSupported ? alternativeIsoCode : 'USD';
+    this.isFiatIsoCodeSupported = _.includes(
+      this.fiatCodes,
+      alternativeIsoCode
+    );
+    this.fiatIsoCode = this.isFiatIsoCodeSupported ? alternativeIsoCode : 'USD';
   }
 
   public getDigitsInfo(coin: string) {
