+++ src/providers/providers.module.ts
@@ -33,7 +33,6 @@ import {
   ElectronProvider,
   EmailNotificationsProvider,
   ErrorsProvider,
-  ExchangeRatesProvider,
   ExternalLinkProvider,
   FCMNG,
   FeedbackProvider,
@@ -137,7 +136,6 @@ import {
     OnGoingProcessProvider,
     PayproProvider,
     PlatformProvider,
-    ExchangeRatesProvider,
     ProfileProvider,
     PopupProvider,
     QRScanner,
